[sw/silicon_creator] Move mock_abs_mmio to device/lib.
Move mock_abs_mmio out of silicon_creator to device/lib/base/testing,
and update tests and mocks to use it from the new location.
