+++ sw/device/lib/base/testing/BUILD
@@ -28,6 +28,20 @@ cc_library(
     ],
 )
 
+cc_library(
+    name = "mock_abs_mmio",
+    testonly = True,
+    hdrs = [
+        "mock_abs_mmio.h",
+    ],
+    defines = ["MOCK_ABS_MMIO"],
+    deps = [
+        ":global_mock",
+        "//sw/device/lib/base:abs_mmio",
+        "@googletest//:gtest",
+    ],
+)
+
 cc_test(
     name = "mock_mmio_test",
     srcs = [
