+++ sw/device/lib/base/testing/meson.build
@@ -12,20 +12,6 @@ sw_lib_testing_bitfield = declare_dependency(
   )
 )
 
-# global mock class
-sw_lib_testing_global_mock = declare_dependency(
-  link_with: static_library(
-    'sw_lib_testing_global_mock',
-    sources: [
-      'global_mock.h'
-    ],
-    dependencies: [
-      sw_vendor_gtest,
-    ],
-    native: true,
-  ),
-)
-
 sw_lib_testing_memory = declare_dependency(
   link_with: static_library(
     'memory',
@@ -35,6 +21,22 @@ sw_lib_testing_memory = declare_dependency(
     c_args: ['-fno-builtin'],
     native: true,
   ),
+)
+
+sw_lib_testing_mock_abs_mmio = declare_dependency(
+  link_with: static_library(
+    'mock_abs_mmio',
+    sources: [
+      'mock_abs_mmio.h',
+    ],
+    dependencies: [
+      sw_vendor_gtest,
+      sw_lib_testing_bitfield,
+    ],
+    native: true,
+    c_args: ['-DMOCK_ABS_MMIO'],
+    cpp_args: ['-DMOCK_ABS_MMIO'],
+  )
 )
 
 sw_lib_testing_hardened = declare_dependency(
