+++ sw/device/silicon_creator/lib/BUILD
@@ -339,7 +339,7 @@ cc_test(
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
         "//sw/device/lib/arch:device",
         "//sw/device/lib/base",
-        "//sw/device/silicon_creator/lib/base:mock_abs_mmio",
+        "//sw/device/lib/base/testing:mock_abs_mmio",
         "//sw/device/silicon_creator/lib/drivers:mock_lifecycle",
         "//sw/device/silicon_creator/lib/drivers:mock_alert",
         "//sw/device/silicon_creator/lib/drivers:mock_otp",
