+++ sw/device/silicon_creator/lib/base/BUILD
@@ -30,22 +30,6 @@ cc_test(
 )
 
 cc_library(
-    name = "mock_abs_mmio",
-    testonly = True,
-    hdrs = [
-        "mock_abs_mmio.h",
-        "//sw/device/lib/base:abs_mmio_h",
-    ],
-    defines = ["MOCK_ABS_MMIO"],
-    deps = [
-        "//sw/device/lib/base/testing",
-        "//sw/device/lib/base/testing:global_mock",
-        "//sw/device/silicon_creator/testing:mask_rom_test",
-        "@googletest//:gtest",
-    ],
-)
-
-cc_library(
     name = "sec_mmio",
     srcs = ["sec_mmio.c"],
     hdrs = ["sec_mmio.h"],
@@ -81,8 +65,8 @@ cc_test(
     ],
     defines = ["OT_OFF_TARGET_TEST"],
     deps = [
-        ":mock_abs_mmio",
         "//sw/device/lib/base",
+        "//sw/device/lib/base/testing:mock_abs_mmio",
         "//sw/device/silicon_creator/lib:error",
         "@googletest//:gtest_main",
     ],
