+++ sw/device/silicon_creator/lib/drivers/BUILD
@@ -101,8 +101,8 @@ cc_test(
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
         "//sw/device/lib/base",
         "//sw/device/lib/base/testing",
+        "//sw/device/lib/base/testing:mock_abs_mmio",
         "//sw/device/silicon_creator/lib:error",
-        "//sw/device/silicon_creator/lib/base:mock_abs_mmio",
         "//sw/device/silicon_creator/lib/base:mock_sec_mmio",
         "//sw/device/silicon_creator/lib/drivers:mock_otp",
         "@googletest//:gtest_main",
@@ -141,8 +141,8 @@ cc_test(
         ":hmac",
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
         "//sw/device/lib/base",
+        "//sw/device/lib/base/testing:mock_abs_mmio",
         "//sw/device/silicon_creator/lib:error",
-        "//sw/device/silicon_creator/lib/base:mock_abs_mmio",
         "@googletest//:gtest_main",
     ],
 )
@@ -186,9 +186,9 @@ cc_test(
         "//hw/ip/keymgr/data:keymgr_regs",
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
         "//sw/device/lib/base/freestanding",
+        "//sw/device/lib/base/testing:mock_abs_mmio",
         "//sw/device/silicon_creator/lib:error",
         "//sw/device/silicon_creator/lib:keymgr_binding",
-        "//sw/device/silicon_creator/lib/base:mock_abs_mmio",
         "//sw/device/silicon_creator/lib/base:mock_sec_mmio",
         "@googletest//:gtest_main",
     ],
@@ -292,8 +292,8 @@ cc_test(
         "//hw/ip/otbn/data:otbn_regs",
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
         "//sw/device/lib/base",
+        "//sw/device/lib/base/testing:mock_abs_mmio",
         "//sw/device/silicon_creator/lib:error",
-        "//sw/device/silicon_creator/lib/base:mock_abs_mmio",
         "//sw/device/silicon_creator/testing:mask_rom_test",
         "@googletest//:gtest_main",
     ],
@@ -365,7 +365,7 @@ cc_test(
         ":pinmux",
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
         "//sw/device/lib/base",
-        "//sw/device/silicon_creator/lib/base:mock_abs_mmio",
+        "//sw/device/lib/base/testing:mock_abs_mmio",
         "@googletest//:gtest_main",
     ],
 )
@@ -390,7 +390,7 @@ cc_test(
         ":retention_sram",
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
         "//sw/device/lib/base",
-        "//sw/device/silicon_creator/lib/base:mock_abs_mmio",
+        "//sw/device/lib/base/testing:mock_abs_mmio",
         "@googletest//:gtest_main",
     ],
 )
@@ -474,8 +474,8 @@ cc_test(
         "//hw/top_earlgrey/ip/rstmgr/data/autogen:rstmgr_regs",
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
         "//sw/device/lib/base",
+        "//sw/device/lib/base/testing:mock_abs_mmio",
         "//sw/device/silicon_creator/lib:error",
-        "//sw/device/silicon_creator/lib/base:mock_abs_mmio",
         "@googletest//:gtest_main",
     ],
 )
@@ -501,7 +501,7 @@ cc_test(
         ":uart",
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
         "//sw/device/lib/base",
-        "//sw/device/silicon_creator/lib/base:mock_abs_mmio",
+        "//sw/device/lib/base/testing:mock_abs_mmio",
         "@googletest//:gtest_main",
     ],
 )
@@ -545,7 +545,7 @@ cc_test(
         "//hw/ip/otp_ctrl/data:otp_ctrl_regs",
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
         "//sw/device/lib/base",
-        "//sw/device/silicon_creator/lib/base:mock_abs_mmio",
+        "//sw/device/lib/base/testing:mock_abs_mmio",
         "//sw/device/silicon_creator/lib/base:mock_sec_mmio",
         "//sw/device/silicon_creator/lib/drivers:mock_otp",
         "@googletest//:gtest_main",
