+++ sw/device/silicon_creator/lib/drivers/otbn_unittest.cc
@@ -7,7 +7,7 @@
 #include <array>
 
 #include "gtest/gtest.h"
-#include "sw/device/silicon_creator/lib/base/mock_abs_mmio.h"
+#include "sw/device/lib/base/testing/mock_abs_mmio.h"
 #include "sw/device/silicon_creator/testing/mask_rom_test.h"
 
 #include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
