+++ sw/device/silicon_creator/lib/drivers/watchdog_unittest.cc
@@ -5,7 +5,7 @@
 #include "sw/device/silicon_creator/lib/drivers/watchdog.h"
 
 #include "gtest/gtest.h"
-#include "sw/device/silicon_creator/lib/base/mock_abs_mmio.h"
+#include "sw/device/lib/base/testing/mock_abs_mmio.h"
 #include "sw/device/silicon_creator/lib/base/mock_sec_mmio.h"
 #include "sw/device/silicon_creator/lib/drivers/lifecycle.h"
 #include "sw/device/silicon_creator/lib/drivers/mock_otp.h"
